Generalize sqlite IO manager logic
