+++ pyproject.toml
@@ -9,7 +9,6 @@ dependencies = [
     "sqlalchemy>=2",
     "dagster>=1.7.15, <1.8.3",
     "dagster-webserver>=1.7.15,<1.9",
-    "dagster-duckdb-pandas>=0.24,<0.25",
     "pandas-gbq>=0.23.1",
     "pydata-google-auth>=1.8.2",
     "jupyterlab>=4.2.3",
